+++ CHANGES.rst
@@ -1,3 +1,47 @@
+v13.0.0 (unreleased)
+--------------------
+
+Incorporate changes from workalendar v13.0.0 (2020-11-13)
+
+- Calendars
+
+  - Add optional school holidays to Netherlands calendar, by @Flix6x (#556).
+  - Add optional carnival to Netherlands calendar.
+
+- Documentation
+
+  - Moving the `contributing.md` file to the `docs/` directory (#573).
+  - Changed from `setup.py` to a nice `setup.cfg` file, thanks @ewjoachim (#576).
+  - Added documentation about class options (#572).
+  - Converted `README.rst` file into `README.md` (#575).
+  - Fixed Pull Request template to reference `setup.cfg` (#587).
+
+- Other changes
+
+  - Switched from Travis CI to Github Actions for CI jobs, thanks to @mgu.
+  - Added support of Python 3.9 (#557).
+  - Changed from `setup.py` to a nice `setup.cfg` file, thanks @ewjoachim (#576).
+  - Use the `setup.cfg` file in the key to cache in `ci.yml` file (#587).
+  - [OBSOLETE] Switched from bionic to focal on Travis CI (we've switched to GH actions after that).
+
+v12.1.0 (2020-10-16)
+
+- New calendars
+
+  - Added Spain regions: Andalusia, Aragon, Castile and León, Castilla-La Mancha, Canary Islands, Extremadura, Galicia, Balearic Islands, La Rioja, Community of Madrid, Murcia, Navarre, Asturias, Basque Country, Cantabria, Valencian Community (#531).
+  - Added all ISO codes for Spain regions - thx @ainarela for your help on this (#531).
+
+- Other changes
+
+  - Refactored Spain test modules (#531).
+  - Fix Catalonia calendar by removing *Sant Juan* day, which does not appear to be an official holiday (#531).
+  - Improve coverage of `workalendar/core.py` module (#546).
+  - Improve coverage for the Netherlands calendar - Queen's Day (#546).
+  - Improve coverage for the Romania calendar - Liberation day (#546).
+  - Improve coverage for the New Zealand calendar (#546).
+  - Added a tox entrypoint to ensure code is Python 3.6+, using ``pyupgrade`` (#566).
+  - Added the pyupgrade tox job to the test suite, amended contributing documentation (#566).
+
 v12.0.0 (unreleased)
 --------------------